feat(models): add BaseState and refactor states
