+++ src/local_newsifier/models/state.py
@@ -1,9 +1,10 @@
 from datetime import datetime, timezone
 from enum import Enum
-from typing import Dict, List, Optional, Any
-from uuid import UUID, uuid4
+from typing import Any, Dict, List, Optional
 
-from pydantic import BaseModel, ConfigDict, Field
+from pydantic import ConfigDict, Field
+
+from .base_state import BaseState, ErrorDetails, extract_error_details
 
 
 class AnalysisStatus(str, Enum):
@@ -33,43 +34,11 @@ class TrackingStatus(str, Enum):
     FAILED = "FAILED"
 
 
-class ErrorDetails(BaseModel):
-    """Details about an error that occurred during processing."""
-
-    task: str
-    type: str
-    message: str
-    traceback_snippet: Optional[str] = None
-    
-    
-def extract_error_details(error: Exception) -> tuple:
-    """Extract error details from an exception, unwrapping ServiceError if needed.
-    
-    Args:
-        error: The exception to extract details from
-        
-    Returns:
-        Tuple of (error_type, message, traceback_snippet)
-    """
-    # Extract original error message if it's a ServiceError, otherwise use as is
-    if hasattr(error, 'original') and error.original:
-        # Unwrap the ServiceError to get the original exception
-        error_type = error.original.__class__.__name__
-        message = str(error.original)
-        traceback_snippet = str(error.original.__traceback__)
-    else:
-        # Use the provided exception as is
-        error_type = error.__class__.__name__
-        message = str(error)
-        traceback_snippet = str(error.__traceback__)
-        
-    return error_type, message, traceback_snippet
 
 
-class NewsAnalysisState(BaseModel):
+class NewsAnalysisState(BaseState):
     """State model for the news analysis pipeline."""
 
-    run_id: UUID = Field(default_factory=uuid4)
     target_url: str
     scraped_at: Optional[datetime] = None
     scraped_text: Optional[str] = None
@@ -81,11 +50,7 @@ class NewsAnalysisState(BaseModel):
     saved_at: Optional[datetime] = None
     save_path: Optional[str] = None
     status: AnalysisStatus = Field(default=AnalysisStatus.INITIALIZED)
-    error_details: Optional[ErrorDetails] = None
     retry_count: int = Field(default=0)
-    created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    last_updated: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    run_logs: List[str] = Field(default_factory=list)
 
     model_config = ConfigDict(
         json_schema_extra={
@@ -99,42 +64,18 @@ class NewsAnalysisState(BaseModel):
         }
     )
 
-    def touch(self) -> None:
-        """Update the last_updated timestamp."""
-        self.last_updated = datetime.now(timezone.utc)
-
-    def add_log(self, message: str) -> None:
-        """Add a log message with timestamp."""
-        timestamp = datetime.now(timezone.utc).isoformat()
-        self.run_logs.append(f"[{timestamp}] {message}")
-        self.touch()
-    
-    def set_error(self, task: str, error: Exception) -> None:
-        """Set error details and update status."""
-        self.error_details = ErrorDetails(
-            task=task,
-            type=error.__class__.__name__,
-            message=str(error),
-            traceback_snippet=str(error.__traceback__),
-        )
-        # Don't change the status - the caller is responsible for setting the appropriate status
-        self.touch()
 
 
-class EntityTrackingState(BaseModel):
+class EntityTrackingState(BaseState):
     """State model for the entity tracking process."""
     
-    run_id: UUID = Field(default_factory=uuid4)
     article_id: int
     content: str
     title: str
     published_at: datetime
     entities: List[Dict[str, Any]] = Field(default_factory=list)
     status: TrackingStatus = Field(default=TrackingStatus.INITIALIZED)
-    error_details: Optional[ErrorDetails] = None
-    created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    last_updated: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    run_logs: List[str] = Field(default_factory=list)
+    failure_status: TrackingStatus = TrackingStatus.FAILED
     
     model_config = ConfigDict(
         json_schema_extra={
@@ -150,43 +91,19 @@ class EntityTrackingState(BaseModel):
         }
     )
     
-    def touch(self) -> None:
-        """Update the last_updated timestamp."""
-        self.last_updated = datetime.now(timezone.utc)
-    
-    def add_log(self, message: str) -> None:
-        """Add a log message with timestamp."""
-        timestamp = datetime.now(timezone.utc).isoformat()
-        self.run_logs.append(f"[{timestamp}] {message}")
-        self.touch()
-    
-    def set_error(self, task: str, error: Exception) -> None:
-        """Set error details and update status."""
-        error_type, message, traceback_snippet = extract_error_details(error)
-        self.error_details = ErrorDetails(
-            task=task,
-            type=error_type,
-            message=message,
-            traceback_snippet=traceback_snippet,
-        )
-        self.status = TrackingStatus.FAILED
-        self.touch()
 
 
-class EntityBatchTrackingState(BaseModel):
+class EntityBatchTrackingState(BaseState):
     """State model for batch entity tracking process."""
     
-    run_id: UUID = Field(default_factory=uuid4)
     status: TrackingStatus = Field(default=TrackingStatus.INITIALIZED)
+    failure_status: TrackingStatus = TrackingStatus.FAILED
     status_filter: str = Field(default="analyzed")
     processed_articles: List[Dict[str, Any]] = Field(default_factory=list)
     total_articles: int = 0
     processed_count: int = 0
     error_count: int = 0
     error_details: Optional[ErrorDetails] = None
-    created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    last_updated: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    run_logs: List[str] = Field(default_factory=list)
     
     model_config = ConfigDict(
         json_schema_extra={
@@ -203,27 +120,6 @@ class EntityBatchTrackingState(BaseModel):
         }
     )
     
-    def touch(self) -> None:
-        """Update the last_updated timestamp."""
-        self.last_updated = datetime.now(timezone.utc)
-    
-    def add_log(self, message: str) -> None:
-        """Add a log message with timestamp."""
-        timestamp = datetime.now(timezone.utc).isoformat()
-        self.run_logs.append(f"[{timestamp}] {message}")
-        self.touch()
-    
-    def set_error(self, task: str, error: Exception) -> None:
-        """Set error details and update status."""
-        error_type, message, traceback_snippet = extract_error_details(error)
-        self.error_details = ErrorDetails(
-            task=task,
-            type=error_type,
-            message=message,
-            traceback_snippet=traceback_snippet,
-        )
-        self.status = TrackingStatus.FAILED
-        self.touch()
     
     def add_processed_article(self, article_data: Dict[str, Any], success: bool = True) -> None:
         """Add a processed article to the state."""
@@ -234,20 +130,17 @@ class EntityBatchTrackingState(BaseModel):
         self.touch()
 
 
-class EntityDashboardState(BaseModel):
+class EntityDashboardState(BaseState):
     """State model for entity dashboard generation."""
     
-    run_id: UUID = Field(default_factory=uuid4)
     status: TrackingStatus = Field(default=TrackingStatus.INITIALIZED)
+    failure_status: TrackingStatus = TrackingStatus.FAILED
     days: int = Field(default=30)
     entity_type: str = Field(default="PERSON")
     start_date: Optional[datetime] = None
     end_date: Optional[datetime] = None
     dashboard_data: Dict[str, Any] = Field(default_factory=dict)
     error_details: Optional[ErrorDetails] = None
-    created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    last_updated: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    run_logs: List[str] = Field(default_factory=list)
     
     model_config = ConfigDict(
         json_schema_extra={
@@ -262,43 +155,19 @@ class EntityDashboardState(BaseModel):
         }
     )
     
-    def touch(self) -> None:
-        """Update the last_updated timestamp."""
-        self.last_updated = datetime.now(timezone.utc)
-    
-    def add_log(self, message: str) -> None:
-        """Add a log message with timestamp."""
-        timestamp = datetime.now(timezone.utc).isoformat()
-        self.run_logs.append(f"[{timestamp}] {message}")
-        self.touch()
-    
-    def set_error(self, task: str, error: Exception) -> None:
-        """Set error details and update status."""
-        error_type, message, traceback_snippet = extract_error_details(error)
-        self.error_details = ErrorDetails(
-            task=task,
-            type=error_type,
-            message=message,
-            traceback_snippet=traceback_snippet,
-        )
-        self.status = TrackingStatus.FAILED
-        self.touch()
 
 
-class EntityRelationshipState(BaseModel):
+class EntityRelationshipState(BaseState):
     """State model for entity relationship analysis."""
     
-    run_id: UUID = Field(default_factory=uuid4)
     status: TrackingStatus = Field(default=TrackingStatus.INITIALIZED)
+    failure_status: TrackingStatus = TrackingStatus.FAILED
     entity_id: int
     days: int = Field(default=30)
     start_date: Optional[datetime] = None
     end_date: Optional[datetime] = None
     relationship_data: Dict[str, Any] = Field(default_factory=dict)
     error_details: Optional[ErrorDetails] = None
-    created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    last_updated: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))
-    run_logs: List[str] = Field(default_factory=list)
     
     model_config = ConfigDict(
         json_schema_extra={
@@ -313,24 +182,3 @@ class EntityRelationshipState(BaseModel):
         }
     )
     
-    def touch(self) -> None:
-        """Update the last_updated timestamp."""
-        self.last_updated = datetime.now(timezone.utc)
-    
-    def add_log(self, message: str) -> None:
-        """Add a log message with timestamp."""
-        timestamp = datetime.now(timezone.utc).isoformat()
-        self.run_logs.append(f"[{timestamp}] {message}")
-        self.touch()
-    
-    def set_error(self, task: str, error: Exception) -> None:
-        """Set error details and update status."""
-        error_type, message, traceback_snippet = extract_error_details(error)
-        self.error_details = ErrorDetails(
-            task=task,
-            type=error_type,
-            message=message,
-            traceback_snippet=traceback_snippet,
-        )
-        self.status = TrackingStatus.FAILED
-        self.touch()
